Better comparison, still need refactoring

+++ zgoubidoo_core/physics/fields/fields_init.py
@@ -4,11 +4,11 @@ import numpy as np
 
 @jit(nopython=True)
 def init_field() -> tuple:
-    field_partials = (np.zeros(3),  # B
-                      np.zeros((3, 3)),  # DB in original Zgoubi
-                      np.zeros((3, 3, 3)),  # DDB
-                      np.zeros((3, 3, 3, 3)),  # D3B
-                      np.zeros((3, 3, 3, 3, 3))  # D4B
+    field_partials = (np.zeros(3, dtype=np.float64),  # B
+                      np.zeros((3, 3), dtype=np.float64),  # DB in original Zgoubi
+                      np.zeros((3, 3, 3), dtype=np.float64),  # DDB
+                      np.zeros((3, 3, 3, 3), dtype=np.float64),  # D3B
+                      np.zeros((3, 3, 3, 3, 3), dtype=np.float64)  # D4B
                       )
     return field_partials
